remove dependency of NumberSignal on EventChannel
